% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Covid-19 shiny application}
\source{
The dataset from the application comes from Ramikripsin coronavirus package
}
\description{
This function will find and load the shiny application from inst/app.
}
\examples{
launch_app()

}
