% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southeastcovid-package.R
\docType{package}
\name{southeastcovid-package}
\alias{southeastcovid}
\alias{southeastcovid-package}
\title{southeastcovid: Launch a Covid-19 cases in Southeast Asia}
\description{
This package contains a shiny dashboard application which shows the development of coronavirus cases in South East Asia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etc5523-2020/r-package-assessment-whysptra}
}

}
\author{
\strong{Maintainer}: Putu Saputra \email{psap0002@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
