% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuebox.R
\name{valuebox}
\alias{valuebox}
\title{Refactor the ValueBoxOutput()}
\arguments{
\item{id}{The character vector of InputId}
}
\description{
This function will refactor the ValueBoxOutput() into a single function
}
\examples{
valuebox("confirmed")
valuebox("recovered")
valuebox("death")


}
