% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueboxprint.R
\name{valueboxprint}
\alias{valueboxprint}
\title{Refactor the renderValueBox()}
\arguments{
\item{input}{The input argument (default)}

\item{df}{A dataset which want to summarise (a reactive dataset from newdate function)}

\item{x}{The number of column from x that contains values}

\item{label}{A character vector subtitle that we want to show in value box ("confirmed, death, recovered")}

\item{colors}{A character vector of the available colors (Valid colors are: red, yellow, aqua, blue, light-blue, green, navy, teal, olive, lime, orange, fuchsia, purple, maroon, black.)}
}
\description{
This function will refactor the renderValuebox into a single function thaht can be called within the app
}
\examples{
valueboxprint(input = input, df = newdate(input)(), x = 7, label = "recovered", colors = "aqua")
}
